LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/registration-updateviacache.https-expected.txt


FAIL register-with-updateViaCache-undefined assert_not_equals: Main script should have updated got disallowed value 1512338030.075291
FAIL register-with-updateViaCache-imports assert_not_equals: Main script should have updated got disallowed value 1512338030.414841
FAIL register-with-updateViaCache-all assert_equals: No new service worker expected null but got object "[object ServiceWorker]"
FAIL register-with-updateViaCache-none assert_not_equals: Main script should have updated got disallowed value 1512338030.773353
PASS register-with-updateViaCache-undefined-then-undefined 
PASS register-with-updateViaCache-undefined-then-imports 
FAIL register-with-updateViaCache-undefined-then-all assert_equals: reg.updateViaCache updated expected "all" but got "imports"
FAIL register-with-updateViaCache-undefined-then-none assert_equals: reg.updateViaCache updated expected "none" but got "imports"
PASS register-with-updateViaCache-imports-then-undefined 
PASS register-with-updateViaCache-imports-then-imports 
FAIL register-with-updateViaCache-imports-then-all assert_equals: reg.updateViaCache updated expected "all" but got "imports"
FAIL register-with-updateViaCache-imports-then-none assert_equals: reg.updateViaCache updated expected "none" but got "imports"
FAIL register-with-updateViaCache-all-then-undefined assert_equals: reg.updateViaCache updated expected "imports" but got "all"
FAIL register-with-updateViaCache-all-then-imports assert_equals: reg.updateViaCache updated expected "imports" but got "all"
PASS register-with-updateViaCache-all-then-all 
FAIL register-with-updateViaCache-all-then-none assert_equals: reg.updateViaCache updated expected "none" but got "all"
FAIL register-with-updateViaCache-none-then-undefined assert_equals: reg.updateViaCache updated expected "imports" but got "none"
FAIL register-with-updateViaCache-none-then-imports assert_equals: reg.updateViaCache updated expected "imports" but got "none"
FAIL register-with-updateViaCache-none-then-all assert_equals: reg.updateViaCache updated expected "all" but got "none"
PASS register-with-updateViaCache-none-then-none 
